+++ core/src/test/java/arez/FifoTaskQueueTest.java
@@ -25,7 +25,8 @@ public class FifoTaskQueueTest
   {
     final FifoTaskQueue queue = new FifoTaskQueue( 20 );
 
-    final Task task = new Task( Arez.context(), "A", ValueUtil::randomString );
+    final ArezContext context = Arez.context();
+    final Task task = context.task( "A", ValueUtil::randomString );
     queue.queueTask( task );
     assertInvariantFailure( () -> queue.queueTask( task ),
                             "Arez-0098: Attempting to queue task named 'A' when task is already queued." );
@@ -41,18 +42,18 @@ public class FifoTaskQueueTest
     assertNull( queue.dequeueTask() );
 
     final ArezContext context = Arez.context();
-    queue.queueTask( new Task( context, "A", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "A", ValueUtil::randomString ) );
     assertEquals( queue.getQueueSize(), 1 );
     assertTrue( queue.hasTasks() );
-    queue.queueTask( new Task( context, "B", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "B", ValueUtil::randomString ) );
     assertEquals( queue.getQueueSize(), 2 );
-    queue.queueTask( new Task( context, "C", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "C", ValueUtil::randomString ) );
     assertEquals( queue.getQueueSize(), 3 );
-    queue.queueTask( new Task( context, "D", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "D", ValueUtil::randomString ) );
     assertEquals( queue.getQueueSize(), 4 );
-    queue.queueTask( new Task( context, "E", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "E", ValueUtil::randomString ) );
     assertEquals( queue.getQueueSize(), 5 );
-    queue.queueTask( new Task( context, "F", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "F", ValueUtil::randomString ) );
     assertEquals( queue.getQueueSize(), 6 );
 
     assertEquals( queue.getBuffer().size(), 6 );
@@ -103,12 +104,12 @@ public class FifoTaskQueueTest
     final FifoTaskQueue queue = new FifoTaskQueue( 20 );
 
     final ArezContext context = Arez.context();
-    queue.queueTask( new Task( context, "A", ValueUtil::randomString ) );
-    queue.queueTask( new Task( context, "B", ValueUtil::randomString ) );
-    queue.queueTask( new Task( context, "C", ValueUtil::randomString ) );
-    queue.queueTask( new Task( context, "D", ValueUtil::randomString ) );
-    queue.queueTask( new Task( context, "E", ValueUtil::randomString ) );
-    queue.queueTask( new Task( context, "F", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "A", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "B", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "C", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "D", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "E", ValueUtil::randomString ) );
+    queue.queueTask( context.task( "F", ValueUtil::randomString ) );
 
     assertEquals( queue.getQueueSize(), 6 );